+++ pkg/runner/operations/apply/apply.go
@@ -10,11 +10,10 @@ import (
 	"github.com/kyverno/chainsaw/pkg/runner/check"
 	"github.com/kyverno/chainsaw/pkg/runner/cleanup"
 	"github.com/kyverno/chainsaw/pkg/runner/logging"
+	"github.com/kyverno/chainsaw/pkg/runner/mutate"
 	"github.com/kyverno/chainsaw/pkg/runner/namespacer"
 	"github.com/kyverno/chainsaw/pkg/runner/operations"
 	"github.com/kyverno/chainsaw/pkg/runner/operations/internal"
-	mapsutils "github.com/kyverno/chainsaw/pkg/utils/maps"
-	"github.com/kyverno/kyverno-json/pkg/engine/mutate"
 	kerrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/types"
@@ -55,62 +54,16 @@ func New(
 	}
 }
 
-func clean(in any) map[string]any {
-	data, ok := in.(map[any]any)
-	if !ok {
-		return nil
-	}
-	out := map[string]any{}
-	for k, v := range data {
-		if c := clean(v); c != nil {
-			out[k.(string)] = c
-		} else {
-			out[k.(string)] = v
-		}
-	}
-	return out
-}
-
 func (o *operation) Exec(ctx context.Context) (err error) {
 	obj := o.base
 	logger := internal.GetLogger(ctx, &obj)
 	defer func() {
 		internal.LogEnd(logger, logging.Apply, err)
 	}()
-	for _, modifier := range o.modifiers {
-		// if a match is specified, skip the check if the resource doesn't match
-		if modifier.Match != nil && modifier.Match.Value != nil {
-			if errs, err := check.Check(ctx, obj.UnstructuredContent(), nil, modifier.Match); err != nil {
-				return err
-			} else if len(errs) != 0 {
-				continue
-			}
-		}
-		merge := modifier.Merge
-		if modifier.Label != nil {
-			merge = &v1alpha1.Any{
-				Value: map[string]any{
-					"metadata": map[string]any{
-						"labels": modifier.Label.Value,
-					},
-				},
-			}
-		} else if modifier.Annotate != nil {
-			merge = &v1alpha1.Any{
-				Value: map[string]any{
-					"metadata": map[string]any{
-						"annotations": modifier.Annotate.Value,
-					},
-				},
-			}
-		}
-		if merge != nil {
-			patch, err := mutate.Mutate(ctx, nil, mutate.Parse(ctx, merge.Value), obj.UnstructuredContent(), o.bindings)
-			if err != nil {
-				return err
-			}
-			obj.SetUnstructuredContent(mapsutils.Merge(obj.UnstructuredContent(), clean(patch)))
-		}
+	if merged, err := mutate.Merge(ctx, obj, o.bindings, o.modifiers...); err != nil {
+		return err
+	} else {
+		obj = merged
 	}
 	if err := internal.ApplyNamespacer(o.namespacer, &obj); err != nil {
 		return err
